+++ tests/conftest.py
@@ -23,3 +23,21 @@ def test_client(app):
 @pytest.fixture(scope="session")
 def get_json_headers():
     return {"Content-Type": "application/json"}
+
+
+@pytest.fixture
+def sample_kakao_data():
+    # NOTE https://developers.kakao.com/docs/latest/ko/kakaologin/rest-api#req-user-info 참고
+    # nickname 만 동의한 상황
+    return {
+        "id": 123456789,
+        "kakao_account": {
+            "profile_nickname_needs_agreement": False,
+            "profile": {"nickname": "홍길동"},
+        },
+        "properties": {
+            "nickname": "홍길동카톡",
+            "custom_field1": "23",
+            "custom_field2": "여",
+        },
+    }
