+++ tests/utils/test_social_login_helper.py
@@ -3,24 +3,6 @@ from app.utils.social_login import KakaoLoginHelper
 from app.core.exceptions import UnexpectedApiResponseException, UnexpectedDataException
 
 
-@pytest.fixture
-def sample_kakao_data():
-    # NOTE https://developers.kakao.com/docs/latest/ko/kakaologin/rest-api#req-user-info 참고
-    # nickname 만 동의한 상황
-    return {
-        "id": 123456789,
-        "kakao_account": {
-            "profile_nickname_needs_agreement": False,
-            "profile": {"nickname": "홍길동"},
-        },
-        "properties": {
-            "nickname": "홍길동카톡",
-            "custom_field1": "23",
-            "custom_field2": "여",
-        },
-    }
-
-
 def test_parsing_wrong_data_should_raise(requests_mock):
     requests_mock.get("https://kapi.kakao.com/v2/user/me", json={"a": "b"}, status_code=200)
     with pytest.raises(UnexpectedDataException):
